requirements.txt
anyio==4.12.1
fastapi==0.135.1
google-adk==1.26.0
google-genai==1.66.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
postgrest==2.28.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
realtime==2.28.0
sse-starlette==3.3.2
starlette==0.52.1
storage3==2.28.0
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
uvicorn==0.41.0
websockets==15.0.1
